requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
abseil-cpp=20211102.0=h27087fc_1
absl-py=1.4.0=pypi_0
accelerate=0.21.0=pypi_0
aiohttp=3.8.1=py39hb9d737c_1
aiosignal=1.3.1=pyhd8ed1ab_0
antlr4-python3-runtime=4.9.3=pypi_0
arrow-cpp=11.0.0=h374c478_1
async-timeout=4.0.2=pyhd8ed1ab_0
attrs=23.1.0=pyh71513ae_1
aws-c-common=0.6.8=h5eee18b_1
aws-c-event-stream=0.1.6=h6a678d5_6
aws-checksums=0.1.11=h5eee18b_2
aws-sdk-cpp=1.8.185=h721c034_1
bert-score=0.3.13=pypi_0
blas=1.0=mkl
boost-cpp=1.73.0=h7f8727e_12
bottleneck=1.3.5=py39h7deecbd_0
braceexpand=0.1.7=pypi_0
brotli-python=1.0.9=py39h5a03fae_7
bzip2=1.0.8=h7f98852_4
c-ares=1.19.0=h5eee18b_0
ca-certificates=2023.7.22=hbcca054_0
certifi=2023.7.22=pyhd8ed1ab_0
cffi=1.15.0=py39h4bc2ebd_0
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.1.6=unix_pyh707e725_0
colorama=0.4.6=pyhd8ed1ab_0
contourpy=1.1.0=pypi_0
cuda=11.6.1=0
cuda-cccl=11.6.55=hf6102b2_0
cuda-command-line-tools=11.6.2=0
cuda-compiler=11.6.2=0
cuda-cudart=11.6.55=he381448_0
cuda-cudart-dev=11.6.55=h42ad0f4_0
cuda-cuobjdump=11.6.124=h2eeebcb_0
cuda-cupti=11.6.124=h86345e5_0
cuda-cuxxfilt=11.6.124=hecbf4f6_0
cuda-driver-dev=11.6.55=0
cuda-gdb=12.2.53=0
cuda-libraries=11.6.1=0
cuda-libraries-dev=11.6.1=0
cuda-memcheck=11.8.86=0
cuda-nsight=12.2.53=0
cuda-nsight-compute=12.2.0=0
cuda-nvcc=11.6.124=hbba6d2d_0
cuda-nvdisasm=12.2.53=0
cuda-nvml-dev=11.6.55=haa9ef22_0
cuda-nvprof=12.2.60=0
cuda-nvprune=11.6.124=he22ec0a_0
cuda-nvrtc=11.6.124=h020bade_0
cuda-nvrtc-dev=11.6.124=h249d397_0
cuda-nvtx=11.6.124=h0630a44_0
cuda-nvvp=12.2.60=0
cuda-runtime=11.6.1=0
cuda-samples=11.6.101=h8efea70_0
cuda-sanitizer-api=12.2.53=0
cuda-toolkit=11.6.1=0
cuda-tools=11.6.1=0
cuda-visual-tools=11.6.1=0
cycler=0.11.0=pypi_0
dataclasses=0.8=pyhc8e2a94_3
datasets=2.13.1=pyhd8ed1ab_0
decord=0.6.0=pypi_0
dill=0.3.6=pyhd8ed1ab_1
einops=0.6.1=pypi_0
einops-exts=0.0.4=pypi_0
evaluate=0.4.0=pypi_0
ffmpeg=4.3=hf484d3e_0
filelock=3.12.2=pyhd8ed1ab_0
fonttools=4.42.0=pypi_0
freetype=2.12.1=h4a9f257_0
frozenlist=1.3.3=py39h5eee18b_0
fsspec=2023.6.0=pyh1a96a4e_0
ftfy=6.1.1=pypi_0
future=0.18.3=pyhd8ed1ab_0
gds-tools=1.7.0.149=0
gflags=2.2.2=he1b5a44_1004
giflib=5.2.1=h5eee18b_3
glog=0.5.0=h48cff8f_0
gmp=6.2.1=h295c915_3
gnutls=3.6.15=he1e5248_0
grpc-cpp=1.48.2=he1ff14a_1
huggingface_hub=0.16.4=pyhd8ed1ab_0
icu=58.2=hf484d3e_1000
idna=3.4=pyhd8ed1ab_0
importlib-metadata=6.8.0=pyha770c72_0
importlib-resources=6.0.1=pypi_0
importlib_metadata=6.8.0=hd8ed1ab_0
intel-openmp=2023.1.0=hdb19cb5_46305
iopath=0.1.10=pypi_0
joblib=1.3.0=pyhd8ed1ab_1
jpeg=9e=h5eee18b_1
kiwisolver=1.4.4=pypi_0
krb5=1.20.1=h143b758_1
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
lerc=3.0=h295c915_0
libboost=1.73.0=h28710b8_12
libbrotlicommon=1.0.9=h166bdaf_7
libbrotlidec=1.0.9=h166bdaf_7
libbrotlienc=1.0.9=h166bdaf_7
libcublas=11.9.2.110=h5e84587_0
libcublas-dev=11.9.2.110=h5c901ab_0
libcufft=10.7.1.112=hf425ae0_0
libcufft-dev=10.7.1.112=ha5ce4c0_0
libcufile=1.7.0.149=0
libcufile-dev=1.7.0.149=0
libcurand=10.3.3.53=0
libcurand-dev=10.3.3.53=0
libcurl=8.1.1=h251f7ec_1
libcusolver=11.3.4.124=h33c3c4e_0
libcusparse=11.7.2.124=h7538f96_0
libcusparse-dev=11.7.2.124=hbbe9722_0
libdeflate=1.17=h5eee18b_0
libedit=3.1.20221030=h5eee18b_0
libev=4.33=h516909a_1
libevent=2.1.12=hdbd6064_1
libffi=3.4.4=h6a678d5_0
libgcc-ng=11.2.0=h1234567_1
libgfortran-ng=13.1.0=h69a702a_0
libgfortran5=13.1.0=h15d22d2_0
libgomp=11.2.0=h1234567_1
libiconv=1.16=h7f8727e_2
libidn2=2.3.4=h5eee18b_0
libnghttp2=1.52.0=h2d74bed_1
libnpp=11.6.3.124=hd2722f0_0
libnpp-dev=11.6.3.124=h3c42840_0
libnvjpeg=11.6.2.124=hd473ad6_0
libnvjpeg-dev=11.6.2.124=hb5906b9_0
libopenblas=0.3.20=pthreads_h78a6416_0
libpng=1.6.39=h5eee18b_0
libprotobuf=3.20.3=he621ea3_0
libssh2=1.10.0=hdbd6064_2
libstdcxx-ng=11.2.0=h1234567_1
libtasn1=4.19.0=h5eee18b_0
libthrift=0.15.0=h1795dd8_2
libtiff=4.5.0=h6a678d5_2
libtmux=0.22.1=pypi_0
libunistring=0.9.10=h27cfd23_0
libwebp=1.2.4=h11a3e52_1
libwebp-base=1.2.4=h5eee18b_1
lightning-utilities=0.9.0=pypi_0
loralib=0.1.2=pypi_0
lz4-c=1.9.4=h6a678d5_0
matplotlib=3.7.2=pypi_0
mkl=2023.1.0=h6d00ec8_46342
mkl-service=2.4.0=py39h5eee18b_1
mkl_fft=1.3.6=py39h417a72b_1
mkl_random=1.2.2=py39h417a72b_1
multidict=6.0.2=py39h5eee18b_0
multiprocess=0.70.14=py39h06a4308_0
ncurses=6.4=h6a678d5_0
nettle=3.7.3=hbbd107a_1
ninja=1.10.2=h06a4308_5
ninja-base=1.10.2=hd09550d_5
nltk=3.8.1=pypi_0
nsight-compute=2023.2.0.16=0
numexpr=2.8.4=py39hc78ab66_1
numpy=1.25.0=py39h5f9d8c6_0
numpy-base=1.25.0=py39hb5e798b_0
omegaconf=2.3.0=pypi_0
open-clip-torch=2.20.0=pypi_0
opencv-python=4.8.0.74=pypi_0
openh264=2.1.1=h4ff587b_0
openssl=3.0.10=h7f8727e_2
orc=1.7.4=hb3bc3d3_1
packaging=23.1=pyhd8ed1ab_0
pandas=1.5.3=py39h417a72b_0
peft=0.5.0.dev0=pypi_0
pillow=9.4.0=py39h6a678d5_0
pip=23.1.2=py39h06a4308_0
portalocker=2.7.0=pypi_0
protobuf=3.20.3=py39h6a678d5_0
psutil=5.9.5=pypi_0
pyarrow=11.0.0=py39h992f0b0_0
pycparser=2.21=pyhd8ed1ab_0
pyparsing=3.0.9=pypi_0
pysocks=1.7.1=pyha2e5f31_6
python=3.9.17=h955ad1f_0
python-dateutil=2.8.2=pyhd8ed1ab_0
python-xxhash=2.0.2=py39h5eee18b_1
python_abi=3.9=2_cp39
pytorch=1.13.1=py3.9_cuda11.6_cudnn8.3.2_0
pytorch-cuda=11.6=h867d48c_1
pytorch-mutex=1.0=cuda
pytz=2023.3=pyhd8ed1ab_0
pyyaml=6.0=py39hb9d737c_4
re2=2022.04.01=h27087fc_0
readline=8.2=h5eee18b_0
regex=2022.7.9=py39h5eee18b_0
requests=2.31.0=pyhd8ed1ab_0
responses=0.18.0=pyhd8ed1ab_0
rouge-score=0.1.2=pypi_0
sacremoses=0.0.53=pyhd8ed1ab_0
safetensors=0.3.1=pypi_0
sentencepiece=0.1.99=pypi_0
setuptools=67.8.0=py39h06a4308_0
six=1.16.0=pyh6c4a22f_0
snappy=1.1.9=h295c915_0
sqlite=3.41.2=h5eee18b_0
tbb=2021.8.0=hdb19cb5_0
timm=0.9.5=pypi_0
tk=8.6.12=h1ccaba5_0
tmuxp=1.28.1=pypi_0
tokenizers=0.13.3=pypi_0
torchaudio=0.13.1=py39_cu116
torchvision=0.14.1=py39_cu116
tqdm=4.65.0=pyhd8ed1ab_1
transformers=4.28.1=pyhd8ed1ab_0
typing-extensions=4.7.1=hd8ed1ab_0
typing_extensions=4.7.1=pyha770c72_0
tzdata=2023c=h04d1e81_0
urllib3=2.0.4=pyhd8ed1ab_0
utf8proc=2.6.1=h27cfd23_0
wcwidth=0.2.6=pypi_0
webdataset=0.2.48=pypi_0
wheel=0.38.4=py39h06a4308_0
xxhash=0.8.0=h7f98852_3
xz=5.4.2=h5eee18b_0
yaml=0.2.5=h7f98852_2
yarl=1.7.2=py39hb9d737c_2
zipp=3.16.2=pyhd8ed1ab_0
zlib=1.2.13=h5eee18b_0
zstd=1.5.5=hc292b87_0
